services/auth/requirements.txt
Flask
Flask-CORS
python-dotenv
requests
PyJWT
Flask-SQLAlchemy